Create component folder and move App into it
